Pastes CRUD

+++ src/engine/src/comment.rs
@@ -2,15 +2,15 @@ use crate::{models::Comment, utils};
 
 impl Comment {
     fn new(
-        paste_id: usize,
-        id: usize,
+        paste_id: i64,
+        id: i64,
         content: String,
         author: Option<String>,
-        created_at: u64,
-        from_row: usize,
-        from_column: usize,
-        to_row: usize,
-        to_column: usize,
+        created_at: i64,
+        from_row: i64,
+        from_column: i64,
+        to_row: i64,
+        to_column: i64,
     ) -> Self {
         Self {
             paste_id,
@@ -27,13 +27,13 @@ impl Comment {
 
     // should this be here? or creation inside paste entirely?
     pub async fn create(
-        paste_id: usize,
+        paste_id: i64,
         content: String,
         author: Option<String>,
-        from_row: usize,
-        from_column: usize,
-        to_row: usize,
-        to_column: usize,
+        from_row: i64,
+        from_column: i64,
+        to_row: i64,
+        to_column: i64,
     ) -> Self {
         Self::new(
             paste_id,
